+++ CHANGELOG.md
@@ -5,6 +5,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Better error message when trying to transform a geometry with infinite bounds.
+
 ## [2.6.1] - 2024-03-07
 
 ### Added